feat(F-417): remove verification service
